Change homepage: image scales size and h1 styling
Changed the size of the homepage image file to 410px x 300px and removed
 the width and height attributes from index.html. Then, I added some css
 attributes that modify img tags. I set width and height to auto, then
I set max-width and max-height to 100%. This causes the image to retain
it's designated place in the html document and to scale with the screen
size of the device used to view it. I also changed the homepage h1 tag
to an inline-flex to make a styled curved outline and to change color
when the user hovers the mouse over it. !!!the h1 tag is inline-flex in
 this commit, I will change it to inline-flex in a moment, sorry!!!