Modify backgrounds and h1 styling all pages
I didn't add the main style.css in the last commit. I modified the h1
tags on all pages and the backgrounds on all pages. I am adding the
updated style.css which was missed in the last commit. Consider this
as the full log comments for the previous commit.